Move registerFunc to testutils
